Move ParameterCountRange to utils
This files is only used in file in utils and has nothing to do with
resolve.
